Fix Makefile and update mailbox.c and transmitter.cpp for precision and linking issues

+++ mailbox.c
@@ -61,7 +61,8 @@ void *mapmem(unsigned base, unsigned size)
    printf("base=0x%x, mem=%p\n", base, mem);
 #endif
    if (mem == MAP_FAILED) {
-      printf("mmap error %d\n", (int)mem);
+      printf("mmap error %ld\n", (intptr_t)mem);
+      //printf("mmap error %d\n", (int)mem);
       exit (-1);
    }
    close(mem_fd);
